Expose ForceFlush and use it in WaitAgentPresence.
