+++ state/presence/pingbatcher.go
@@ -179,6 +179,24 @@ func (pb *PingBatcher) Ping(modelUUID string, slot int64, fieldKey string, field
 	}
 }
 
+// ForceFlush immediately flushes the current state to the database.
+// This should generally only be called from testing code, everyone else can
+// generally wait the usual wait for updates to be flushed naturally.
+func (pb *PingBatcher) ForceFlush() error {
+	request := make(chan struct{})
+	select {
+	case pb.flushChan <- request:
+		select {
+		case <-request:
+			return nil
+		case <-pb.tomb.Dying():
+			return pb.tomb.Err()
+		}
+	case <-pb.tomb.Dying():
+		return pb.tomb.Err()
+	}
+}
+
 // handlePing is where we actually update our internal structures after we
 // get a ping request.
 func (pb *PingBatcher) handlePing(ping singlePing) {
